food/avatar/requirements.txt
fastapi
numpy
redis
pillow
uvicorn